main copied

+++ app/Http/Controllers/CategoryController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Models\Product;
 use App\Models\Category;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
@@ -38,6 +39,7 @@ class CategoryController extends Controller
             'status' => 'required|in:1,3', 
             'icon_file' => 'nullable|max:2048', 
             'background_image' => 'nullable|max:2048', 
+            'banner_image' => 'nullable|max:2048', 
         ]);
     
         $category = new Category();
@@ -53,6 +55,12 @@ class CategoryController extends Controller
         if ($request->hasFile('background_image')) {
             $path = $request->file('background_image')->store('public/backgrounds');
             $category->background_image = str_replace('public/', '', $path);
+        }
+
+        // Handle banner_image upload
+        if ($request->hasFile('banner_image')) {
+            $path = $request->file('banner_image')->store('public/banners');
+            $category->banner_image = str_replace('public/', '', $path);
         }
     
         $category->created_by = Auth::id(); 
@@ -89,11 +97,23 @@ class CategoryController extends Controller
             'status' => 'required|in:1,3',
             'icon_file' => 'nullable|file|image|max:2048',
             'background_image' => 'nullable|file|image|max:2048',
+            'banner_image' => 'nullable|file|image|max:2048',
         ]);
     
         $category = Category::findOrFail($id);
         $category->fill($validatedData);
     
+        // remove icon file
+        if($request->input('icon_file_remove')){
+            $category->icon_file = null;
+        }
+        if($request->input('background_image_remove')){
+            $category->background_image = null;
+        }
+        if($request->input('banner_image_remove')){
+            $category->banner_image = null;
+        }
+
         // Handle icon_file upload
         if ($request->hasFile('icon_file')) {
             $path = $request->file('icon_file')->store('public/icons');
@@ -104,6 +124,12 @@ class CategoryController extends Controller
         if ($request->hasFile('background_image')) {
             $path = $request->file('background_image')->store('public/backgrounds');
             $category->background_image = str_replace('public/', '', $path);
+        }
+
+        // Handle banner_image upload
+        if ($request->hasFile('banner_image')) {
+            $path = $request->file('banner_image')->store('public/banners');
+            $category->banner_image = str_replace('public/', '', $path);
         }
     
         $category->updated_by = Auth::id();
@@ -121,9 +147,16 @@ class CategoryController extends Controller
     {
         try {
             $category = Category::findOrFail($id);
-            $category->delete();
-    
-            return redirect()->route('category.list')->with('success', 'Category deleted successfully.');
+            if($category){
+                $products = Product::where('category_id', $id)->get();
+                if(count($products)){
+                    return redirect()->route('category.list')->with('error', 'There are products related to this category. First delete or update their category');
+                }
+                else{
+                    $category->delete();
+                    return redirect()->route('category.list')->with('success', 'Category deleted successfully.');
+                }
+            }
         } catch (\Exception $e) {
             return back()->withErrors(['error' => $e->getMessage()]);
         }
